# services/api/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Gamics API
  version: 0.9.0
  description: >-
    Gamics player, competition, blind match result verification, Gamics result review, rankings, M-Pesa payments,
    notification, identity and role-based operations API.
servers:
  - url: http://localhost:8080
paths:
...
  /v1/me/game-accounts:
    get:
      operationId: listMyGameAccounts
      tags:
        - Game accounts
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Connected game accounts.
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/GameAccount'
    post:
      operationId: createGameAccount
      tags:
        - Game accounts
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GameAccountInput'
      responses:
        '201':
          description: Connected game account.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GameAccount'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    GameAccountInput:
      type: object
      required:
        - gameId
        - platform
        - inGameName
      properties:
        gameId:
          type: string
        platform:
          type: string
          enum:
            - android
            - ios
        inGameName:
          type: string
          minLength: 2
          maxLength: 80
        publisherPlayerId:
          type:
            - string
            - 'null'
    GameAccount:
      allOf:
        - $ref: '#/components/schemas/GameAccountInput'
        - type: object
          required:
            - id
            - verificationStatus
            - verificationMethod
            - publisherVerified
            - createdAt
            - updatedAt
          properties:
            id:
              type: string
              format: uuid
            verificationStatus:
              type: string
              enum:
                - unverified
                - pending
                - verified
                - rejected
            verificationMethod:
              type:
                - string
                - 'null'
              enum:
                - manual_evidence
                - publisher_api
                - null
              description: How the account was verified; null while it is not verified.
            publisherVerified:
              type: boolean
              description: True only when verification came from an integrated publisher API; a Gamics evidence review remains false.
            verifiedAt:
              type:
                - string
                - 'null'
              format: date-time
            createdAt:
              type: string
              format: date-time
            updatedAt:
              type: string
              format: date-time
...